move values to different files
